Update color help and index.css
This will likely need to be fixed up in the future. I don't like having
to include the :not selectors in the main CSS file just for the helper
file to work, if it happens to be included. I can fix this problem by
including !important inside the helper CSS file, but that may cascade
into larger problems later, when more helpers come into play.

I'll try and reduce the specificity of these elements in the main file
where having a specificity as high as is present is unnecessary.
Ideally, the helper file, being included after the main file, would
simply override the main file, but this is not the case because of the
specificity. I could also artificially up the specificity of the helper
file's style rules, but that could also result in unexpected annoyances
and issues further down the line.